feat: give other block entities access to display the new recipe item requirements
