+++ src/main/java/com/tcm/MineTale/block/workbenches/menu/AbstractWorkbenchContainerMenu.java
@@ -1,7 +1,6 @@
 package com.tcm.MineTale.block.workbenches.menu;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
 
 import org.jspecify.annotations.Nullable;
